Unhook all plthook hooks when they are not longer needed
Works for the majority of the hooks once mono is initialising since we know by this point, most hooks can be cleaned up. This is done by doing a plthook replace, but back to the original. The only mandatory hook is WriteFile, but it's unavoidable to controll logging with the option to disable unity logs