Remove the boot config customiser feature
It was deemed unecessary upon learning that the file itself isn't the only way to change these. It's possible to change them through command line arguments by prepanding the key with "--" and having the next argument be its value. All that will be needed is to amend the bug fables internal docs to list all the known arguments, but as far as the file is concerned, VRL doesn't need this feature