Add an internal EnumPatcher service
Works on every enums and dynamically determines the ids to reserve based on the biggest existing one. The method was inspired by SRML, but modified for simplicity (no need to use a transpiler).

It primarily involves changing the internal mono cache of an enum values and names. Every addition will cause the cache to be invalidated so consumers can see the changes immediately upon adding the enum value