Change the bootstrap unity logging logic
The setting is no more. It's instead bound to whether or not the UNITY category has trace enabled.

If it's not, the managed loader side will instead use a managed unity log callback that will log with the appropriate levels so we can now filter unity managed logs. This is a compromise between having not having spammy and all trace logs vs still having the option in case of debugging since some logs are exclusive to native logs, but they are very verbose and not useful for most users.